componentization adjustments
